Mind/pyproject.toml
[project]
name = "mind"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "cn2an>=0.5.23",
    "fast-langdetect>=0.3.1",
    "fastapi[standard]>=0.115.2",
    "ffmpeg-python>=0.2.0",
    "funasr==1.0.27",
    "g2p-en>=2.1.0",
    "gradio<5",
    "jieba>=0.42.1",
    "jieba-fast>=0.53",
    "librosa==0.10.2",
    "llama-cpp-python>=0.3.15",
    "matplotlib>=3.10.5",
    "numpy<2.0",
    "peft>=0.17.0",
    "pyopenjtalk>=0.4.1",
    "pypinyin>=0.55.0",
    "pytorch-lightning>=2.4",
    "split-lang>=2.1.0",
    "torchaudio>=2.8.0",
    "torchmetrics<=1.5",
    "wordsegment>=1.3.1",
    "x-transformers>=2.7.0",
]

[dependency-groups]
test = [
    "pytest>=8.4.1",
    "pytest-asyncio>=1.1.0",
    "pytest-tornasync>=0.6.0.post2",
]
